Most logic moved out of main.go and into daemon.go
